feat: refactor app structure by adding event handling and UI management
